+++ bugsnag-android-core/src/main/java/com/bugsnag/android/AppDataCollector.kt
@@ -17,6 +17,7 @@ internal class AppDataCollector(
     private val sessionTracker: SessionTracker,
     private val activityManager: ActivityManager?,
     private val launchCrashTracker: LaunchCrashTracker,
+    private val contextState: ContextState,
     private val logger: Logger
 ) {
     var codeBundleId: String? = null
@@ -46,7 +47,7 @@ internal class AppDataCollector(
     fun getAppDataMetadata(): MutableMap<String, Any?> {
         val map = HashMap<String, Any?>()
         map["name"] = appName
-        map["activeScreen"] = getActiveScreenClass()
+        map["activeScreen"] = contextState.getContext()
         map["memoryUsage"] = getMemoryUsage()
         map["lowMemory"] = isLowMemory()
 
@@ -55,8 +56,6 @@ internal class AppDataCollector(
         }
         return map
     }
-
-    fun getActiveScreenClass(): String? = sessionTracker.contextActivity
 
     /**
      * Get the actual memory used by the VM (which may not be the total used