+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 * Fix rare NullPointerExceptions from ConnectivityManager
   [1311](https://github.com/bugsnag/bugsnag-android/pull/1311)
 
+* Respect manual setting of context
+  [#1310](https://github.com/bugsnag/bugsnag-android/pull/1310)
+
 ## 5.9.5 (2021-06-25)
 
 * Unity: Properly handle ANRs after multiple calls to autoNotify and autoDetectAnrs